% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezec-package.r
\docType{package}
\name{ezec}
\alias{ezec-package}
\alias{ezec}
\title{The \pkg{ezec} package for easy EC calculation.}
\description{
Because fungicide resistance is an important phenotypic trait for fungi and oomycetes, it is necessary to have a standardized method of statistically analyzing the Effective Concentration (EC) values. This package is designed for those who are not terribly familiar with R to be able to analyze and plot an entire set of isolates using the 'drc' package.
}
\section{Introduction}{
 The package \pkg{ezec} is not a revolutionary work. It
  simply is a wrapper for the \pkg{drc} package that makes life a little
  easier when it comes to calculating a simple EC 50. The main function of
  the package is \code{\link{EC_table}}. This function will do as it says and
  automatically produce a table to EC values for each isolate in your sample.
}

\section{Data format}{
 Data is expected to exist in a table with at least
  three columns: \itemize{ \item Sample ID \item Dosage \item Response value
  (Growth) } Any other columns in your data are optional. An example data set
  is \code{\link{dummydata}}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/grunwaldlab/ezec}
  \item Report bugs at \url{https://github.com/grunwaldlab/ezec/issues}
}

}
\author{
\strong{Maintainer}: Zhian N. Kamvar \email{zkamvar@gmail.com}

Other contributors:
\itemize{
  \item Niklaus J. Grunwald \email{grunwaln@science.oregonstate.edu} [thesis advisor, contributor]
}

}
