% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ec.R
\name{get_EC}
\alias{get_EC}
\title{Function to get EC values given a model.}
\usage{
get_EC(x, response = c(10, 50, 90), disp = TRUE)
}
\arguments{
\item{x}{a model generated from the drm function of the drc package}

\item{response}{a numeric vector specifying at what response levels to generate an EC value. Defaults to 10, 50 and 90 (percent).}

\item{disp}{display the results?}
}
\value{
a data frame with the Estimate and standard error for each response
 value.
}
\description{
Function to get EC values given a model.
}
\details{
This function is a wrapper for \code{\link[drc]{ED}}.
}
\author{
Zhian N. Kamvar
}
\keyword{internal}

